requirements.txt
# ======================================
# Hindi ASR Pipeline - Dependencies
# ======================================

# --- Core ML / ASR ---
transformers>=4.36.0
datasets>=2.16.0,<3.0.0
torch>=2.0.0
torchaudio>=2.0.0
accelerate>=0.26.0

# --- Audio Processing ---
librosa>=0.10.0
soundfile>=0.12.0

# --- ASR Evaluation ---
jiwer>=3.0.0

# --- Data Handling ---
pandas>=2.0.0
numpy>=1.24.0

# --- HTTP / Downloads ---
requests>=2.31.0
tqdm>=4.65.0

# --- Environment ---
python-dotenv>=1.0.0

# --- Hindi NLP ---
regex>=2023.0
